requirements.txt
boto3>=1.34
pandas>=2.0
numpy>=1.26
pyarrow>=14.0
python-dotenv>=1.0
pytest>=7.0
faker>=20.0